requirements_dev.txt
-r requirements.txt
django-debug-toolbar==1.4
sqlparse==0.1.16
